Redundant redundancy made less so. New functionality built. New functionality planned.
SuperAdminController has always been a mess in dealing with... having to
recycle some code over and over. Got tired of it. So? I burned it to the
ground and made a special function to handle it.

In other areas, been restructuring much of the framework to make it much
more sensible and predictable, especially for naming conventions. Yay
recycling names!

Also... forms is good to go, just need some changes as to the needs.

Even better though is that finally decided to work on API before
tackling form processing.

In more fun news, had a couple of fun trial runs of the functionality
that already exists. Nothing but positive reviews.